Refactor: #197: category  - deleteComment() 예외처리분류

+++ backend/tiggle-root/tiggle/src/main/kotlin/com/side/tiggle/domain/comment/service/CommentServiceImpl.kt
@@ -84,8 +84,12 @@ class CommentServiceImpl(
 
     override fun deleteComment(memberId: Long, commentId: Long) {
         val comment = commentRepository.findById(commentId)
-            .filter { it.senderId == memberId }
             .orElseThrow { NotFoundException() }
+
+        if (comment.senderId != memberId) {
+            throw NotAuthorizedException()
+        }
+
         commentRepository.delete(comment)
     }
